Refactoring / extract data models from LandscapeController.cs
