+++ Assets/Scripts/Controllers/AlderController.cs
@@ -1,4 +1,4 @@
-﻿using System.Collections;
+using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using System;
@@ -92,8 +92,8 @@ public class AlderController : TreeController {
 		transpiration = newTranspiration;					// -- Randomize between trees?
 
 		if (leafCarbon < maxLeafCarbon)
-//			alphaCutoff = 1f - MapValue (leafCarbon, 0f, maxLeafCarbon, 0f, 0.9f);
-			alphaCutoff = 1f - MapValue (leafCarbon, 0f, maxLeafCarbon, 0f, 0.6f);
+//			alphaCutoff = 1f - MathUtils.MapValue(leafCarbon, 0f, maxLeafCarbon, 0f, 0.9f);
+			alphaCutoff = 1f - MathUtils.MapValue(leafCarbon, 0f, maxLeafCarbon, 0f, 0.6f);
 		else
 			alphaCutoff = 0.1f;
