+++ Assets/Scripts/Controllers/BroadleafController.cs
@@ -1,4 +1,4 @@
-﻿using System.Collections;
+using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
@@ -100,8 +100,8 @@ public class BroadleafController : TreeController {
 		transpiration = newTranspiration;					// -- Randomize between trees?
 
 		if (leafCarbon < maxLeafCarbon)
-//			alphaCutoff = 1f - MapValue (leafCarbon, 0f, maxLeafCarbon, 0f, 0.9f);
-			alphaCutoff = 1f - MapValue (leafCarbon, 0f, maxLeafCarbon, 0f, 0.6f);
+//			alphaCutoff = 1f - MathUtils.MapValue(leafCarbon, 0f, maxLeafCarbon, 0f, 0.9f);
+			alphaCutoff = 1f - MathUtils.MapValue(leafCarbon, 0f, maxLeafCarbon, 0f, 0.6f);
 		else
 			alphaCutoff = 0.1f;
